pyproject.toml
[project]
name = "quali-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=45.0.0",
    "fastapi[all]>=0.134.0",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyjwt[crypto]>=2.9.0",
    "python-ulid>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.47",
    "sqlmodel>=0.0.37",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.4",
    "testcontainers[postgres]>=4.14.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that require a real PostgreSQL container (skip with -m 'not integration')",
]
